We agreed to move these functions back into diracx-db utils module
