+++ diracx-routers/src/diracx/routers/jobs/submission.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 import logging
-from asyncio import TaskGroup
-from copy import deepcopy
 from datetime import datetime, timezone
 from http import HTTPStatus
 from typing import Annotated
@@ -15,6 +13,7 @@ from diracx.core.models import (
     JobStatus,
 )
 from diracx.db.sql.job_logging.db import JobLoggingRecord
+from diracx.db.sql.utils.job import JobSubmissionSpec, submit_jobs_jdl
 
 from ..dependencies import (
     JobDB,
@@ -27,15 +26,6 @@ from .access_policies import ActionType, CheckWMSPolicyCallable
 logger = logging.getLogger(__name__)
 
 router = DiracxRouter()
-
-
-class JobSubmissionSpec(BaseModel):
-    jdl: str
-    owner: str
-    owner_group: str
-    initial_status: str
-    initial_minor_status: str
-    vo: str
 
 
 class InsertedJob(TypedDict):
@@ -76,109 +66,6 @@ StdOutput = std.out;"""
         "value": ["""Arguments = "jobDescription.xml -o LogLevel=INFO"""]
     },
 }
-
-
-async def submit_jobs_jdl(jobs: list[JobSubmissionSpec], job_db: JobDB):
-    from DIRAC.Core.Utilities.ClassAd.ClassAdLight import ClassAd
-    from DIRAC.Core.Utilities.ReturnValues import returnValueOrRaise
-    from DIRAC.WorkloadManagementSystem.DB.JobDBUtils import (
-        checkAndAddOwner,
-        createJDLWithInitialStatus,
-    )
-
-    jobs_to_insert = {}
-    jdls_to_update = {}
-    inputdata_to_insert = {}
-    original_jdls = []
-
-    # generate the jobIDs first
-    # TODO: should ForgivingTaskGroup be used?
-    async with TaskGroup() as tg:
-        for job in jobs:
-            original_jdl = deepcopy(job.jdl)
-            jobManifest = returnValueOrRaise(
-                checkAndAddOwner(original_jdl, job.owner, job.owner_group)
-            )
-
-            # Fix possible lack of brackets
-            if original_jdl.strip()[0] != "[":
-                original_jdl = f"[{original_jdl}]"
-
-            original_jdls.append(
-                (
-                    original_jdl,
-                    jobManifest,
-                    tg.create_task(job_db.create_job(original_jdl)),
-                )
-            )
-
-    async with TaskGroup() as tg:
-        for job, (original_jdl, jobManifest_, job_id_task) in zip(jobs, original_jdls):
-            job_id = job_id_task.result()
-            job_attrs = {
-                "JobID": job_id,
-                "LastUpdateTime": datetime.now(tz=timezone.utc),
-                "SubmissionTime": datetime.now(tz=timezone.utc),
-                "Owner": job.owner,
-                "OwnerGroup": job.owner_group,
-                "VO": job.vo,
-            }
-
-            jobManifest_.setOption("JobID", job_id)
-
-            # 2.- Check JDL and Prepare DIRAC JDL
-            jobJDL = jobManifest_.dumpAsJDL()
-
-            # Replace the JobID placeholder if any
-            if jobJDL.find("%j") != -1:
-                jobJDL = jobJDL.replace("%j", str(job_id))
-
-            class_ad_job = ClassAd(jobJDL)
-
-            class_ad_req = ClassAd("[]")
-            if not class_ad_job.isOK():
-                # Rollback the entire transaction
-                raise ValueError(f"Error in JDL syntax for job JDL: {original_jdl}")
-            # TODO: check if that is actually true
-            if class_ad_job.lookupAttribute("Parameters"):
-                raise NotImplementedError("Parameters in the JDL are not supported")
-
-            # TODO is this even needed?
-            class_ad_job.insertAttributeInt("JobID", job_id)
-
-            await job_db.checkAndPrepareJob(
-                job_id,
-                class_ad_job,
-                class_ad_req,
-                job.owner,
-                job.owner_group,
-                job_attrs,
-                job.vo,
-            )
-            jobJDL = createJDLWithInitialStatus(
-                class_ad_job,
-                class_ad_req,
-                job_db.jdl2DBParameters,
-                job_attrs,
-                job.initial_status,
-                job.initial_minor_status,
-                modern=True,
-            )
-
-            jobs_to_insert[job_id] = job_attrs
-            jdls_to_update[job_id] = jobJDL
-
-            if class_ad_job.lookupAttribute("InputData"):
-                inputData = class_ad_job.getListFromExpression("InputData")
-                inputdata_to_insert[job_id] = [lfn for lfn in inputData if lfn]
-
-        tg.create_task(job_db.update_job_jdls(jdls_to_update))
-        tg.create_task(job_db.insert_job_attributes(jobs_to_insert))
-
-        if inputdata_to_insert:
-            tg.create_task(job_db.insert_input_data(inputdata_to_insert))
-
-    return jobs_to_insert.keys()
 
 
 @router.post("/jdl")
